fixed level_select.py, added exe file

+++ output/main/_internal/source/pan_zoom_sprites/pan_zoom_planet_classes/pan_zoom_planet_params.py
@@ -1,6 +1,5 @@
 import pygame
 
-# from source.database.database_access import create_connection, get_database_file_path
 from source.gui.event_text import event_text
 from source.multimedia_library.images import get_image
 from source.multimedia_library.sounds import sounds
@@ -13,9 +12,7 @@ class PanZoomPlanetParams:
     def __init__(self, kwargs):
         self._image_name_small = kwargs.get("image_name_small")
         self._atmosphere_name = kwargs.get("atmosphere_name")
-
         self.image_name_big = kwargs.get("image_name_big")
-
         self.orbit_radius = 0
         self.font_size = kwargs.get('font_size', 20)
         self.font = kwargs.get('font', pygame.font.SysFont(global_params.font_name, self.font_size))
@@ -47,19 +44,11 @@ class PanZoomPlanetParams:
             self.gif = self._atmosphere_name
             self.setup_gif_handler()
         else:
-            self.gif_handler.kill()
-
-        # # dirty hack to make shure athmosphere_raw gets updated
-        # if value != "":
-        #     if self.has_atmosphere == 1:
-        #         self.atmosphere_raw = images[pictures_path]["atmospheres"][value]
-        #         self.atmosphere = self.atmosphere_raw
-        # else:
-        #     self.atmosphere_raw = None
-        #     self.atmosphere = self.atmosphere_raw
+            if hasattr(self, "gif_handler"):
+                if self.gif_handler:
+                    self.gif_handler.kill()
 
     def set_planet_name(self):
-
         planets_with_same_orbit_object_id = [i for i in sprite_groups.planets if
                                              i.orbit_object_id == self.orbit_object_id]
         sorted_planets = sorted(planets_with_same_orbit_object_id, key=lambda planet: planet.orbit_distance)
@@ -77,8 +66,6 @@ class PanZoomPlanetParams:
         """
         sets the text used for the info_panel
         """
-        # if self.parent.build_menu_visible: return
-        # self.parent.info_panel.visible = True
 
         if self.explored:
             gen_text = info_panel_text_generator.create_info_panel_planet_text(self)
@@ -92,38 +79,6 @@ class PanZoomPlanetParams:
         self.parent.info_panel.set_text(self.info_text)
         self.parent.info_panel.set_planet_image(self.image_raw)
         return
-        # # print (gen_text)
-        #
-        # if self.explored:
-        #     text = self.info_text_raw
-        #     self.info_text = text
-        #     self.parent.info_panel.set_planet_image(self.image_raw)
-        # else:
-        #     text = "unknown planet" + ":\n\n"
-        #     text += "resources: ???\n"
-        #     text += "energy: ???\n"
-        #
-        # text += f"orbit_object_id:{self.orbit_object_id}\n"
-        # text += f"orbit_object:{self.orbit_object}\n"
-        # if self.orbit_angle != None and self.orbit_angle != "None":
-        #     text += f"orbit_angle:{int(self.orbit_angle)}\n"
-        # else:
-        #     text += f"orbit_angle:None\n"
-        # text += f"orbit_distance:{int(self.orbit_distance)}\n"
-        # text += f"x,y:{int(self.world_x), int(self.world_y)}\n"
-        # text += f"_x,_y:{int(self.screen_x), int(self.screen_y)}\n"
-        # text += f"smiley hidden:{self.smiley_button._hidden}\n, smiley disabled:{self.smiley_button._disabled}\n"
-        # text += f"smiley.screen_x:{self.smiley_button.screen_x}, smiley_y:{self.smiley_button.screen_y}\n"
-        # text += f"smiley.x:{self.smiley_button.x}, smiley.y:{self.smiley_button.y}\n"
-        #
-        # self.info_text = text
-        # try:
-        #     self.parent.info_panel.set_text(self.info_text)
-        #     self.parent.info_panel.set_planet_image(self.image_raw)
-        # except AttributeError as e:
-        #     print("set_info_text(self):", e)
-        #
-        # self.parent.info_panel.set_planet_image(self.image_raw)
 
     def get_explored(self):
         """
@@ -141,6 +96,7 @@ class PanZoomPlanetParams:
         self.parent.set_selected_planet(self)
         if not self in self.parent.explored_planets:
             self.parent.add_explored_planet(self)
+
         self.explored = True
         self.show_overview_button()
         self.string = self.name
